Move ERROR to utils
